Adding a Property to a Class
Let's do it! Say public then {dollar_sign}, then the name of your property. In this case, our name is actually name.
Set data on the name property.

Now we can read the data from that property using the same syntax.

Try it out! It's working great.

Does this feel familiar? It works exactly like an array, except instead of the square-bracket syntax, we use this dash, greater-than syntax, let's call that an 'arrow'.

The real difference between an object and an array is that an object has a class that defines all possible properties that it can hold, instead of being able to store any random keys you dream up, like an array.

TIP Technically, you can set a property (e.g. {dollar_sign}myShip->weight = 100) on an object without creating a property for it in the class. But this is frowned upon, and not seen commonly in higher-quality code. So don't do it!

Classes: Data Structure Docs

And with this simple difference, you get a huge benefit. A class is like programmer documentation. If I give you an array, you have no idea what data is on it. You can var_dump it to see, then just hope that it'll always have the same keys in the future.

But if I hand you a Ship object, you know something about it. You know that it will always have a name property, and it'll never have email or phone properties: because they're not in the class. The class gives us a skeleton and some rules we know it'll follow.

Now let's talk about something even better than this: methods.