Arrays and objects have a lot in common: arrays have keys while objects have properties, but both store data.
Objects have one other big, incredible, mind-blowing advantage: methods. A method is a fancy word for a function, and we're all comfortable with those. The difference is that a method lives inside of a class.

To create a method, start with public function. After this, it looks like a normal function - let's call ours sayHello, and add the normal parenthesis and curly braces stuff. Inside, we can do anything - so let's echo Hello.

To call the method, let's use our object 'arrow' syntax. Let's add an <hr/> so I don't confuse myself, then {dollar_sign}myShip, then arrow, then sayHello().

The only difference between accessing a property and a method is the open parenthesis and close parenthesis, so don't forget that. And a method is just like a traditional function, except it lives inside of the class, so you have to call it on the object. And when we refresh, success!

This is kind of amazing because we have little packages of data that can now perform actions and do things through methods. Arrays can't do anything like that.