Referencing Properties from Inside a Class
So let's add another - a getName() method. Remember how functions can return a value? Sure, methods can do that too. Let's not actually return the name of this ship yet, let's fake it.
Down below, we call it the exact same way, except we can echo what it returns.
Now refresh! There's our fake name. Of course, what I really want to do is return the name of the ship that I'm calling this method on. I know, not the most interesting function, but it gives us a new problem. When we have the  variable, we know how to access a property - just use the arrow syntax on it. But when we're inside of the class, how can we get the value of the name property?