When we're inside of the class, how can we get the value of the name property?
The answer is with a very special variable called {dollar_sign}this.

Here's the rule: when you're inside of a method, you magically have access to a variable called . And  is whatever Ship object that we're calling the method on, in our case our favorite {dollar_sign}myShip object whose name is 'Jedi Starship'. In all cases, the variable is called {dollar_sign}this, that's just what the PHP Jedi elders decided this magic name should be.

When we refresh, hey - there's our ship's real name