Our two methods are simple. So now let's add something useful. We'll be printing out the details of our ships all over the place, so I need a way to see the weaponPower, the strength and the jediFactor all at once - like a little summary. One handy way to do this is with a new method that creates that summary string and returns it to us.
Make a new function getNameAndSpecs(). Let's use the nice sprintf function with %s wildcards for the ship's name, followed by the weaponPower, jediFactor and strength. Now, we need to pass it what to fill in for those %s placeholders. To reference the name, use the magic {dollar_sign}this variable: {dollar_sign}this->name. Do the same thing for weaponPower, jediFactor and strength.

And of course, make sure you have a return statement in front of all of this! Let's enjoy our hard work by echo'ing the method below: , arrow, getNameAndSpecs then the open and close parenthesis so PHP knows this is a method, not a property by that name.

Ready to try it? Refresh! There's our weird-little summary. We can use this across our app, and if we ever want to change how it looks, we only need to update one spot.