This object-oriented, or OO, stuff gets really fun once we have multiple objects. Afterall, it takes at least 2 ships to start a battle.
But first, let's summarize all of this printing stuff into a normal, traditional, flat function called printShipSummary(). It'll take in a Ship object as an argument, which I'll call {dollar_sign}someShip. Now, copy all the echo stuff into the function.

The argument to the function could be called anything. Since I chose to call it {dollar_sign}someShip, all of the {dollar_sign}myShip variables below need to be updated to {dollar_sign}someShip also. This is just classic behavior of functions - nothing special.

Ok, saving time!

Back at the bottom of the file, call this like any traditional function, and pass it the {dollar_sign}myShip variable, which we know is a Ship object.

So we're throwing around some objects, but this is just normal, flat, procedural programming. When we refresh it's exactly the same.