Since the goal of our app is to let two ship's fight each other, things are getting interesting. For example, we could fight {dollar_sign}myShip against {dollar_sign}otherShip and see who comes out as the winner.
But first, let's imagine we want to know whose strength is higher. Of course, we could just write an if statement down here and manually check {dollar_sign}myShip's strength against {dollar_sign}otherShip.

But we could also add a new method inside of the Ship class itself. Let's create a new method that'll tells us if one Ship's strength is greater than anothers. We'll call it: doesGivenShipHaveMoreStrength(). And of course, it needs a Ship object as an argument.

So just like with our printShipSummary() function, when we call this function, we'll pass it a Ship object. What I want to do is compare the Ship object being passed to whatever Ship object we're calling this method on.