Inside of the doesGivenShipHaveMoreStrength(), the magic {dollar_sign}this will refer to the {dollar_sign}myShip object, the one whose name is Jedi Starship. So all we need to do is return {dollar_sign}givenShip->strength greater than my strength.
Ok, let's try it! When we refresh, we see that the Imperial Shuttle has more strength. And that makes sense: the Imperial Shuttle has 50 compared with 0 for , because it's using the default value.

Let me add another separator and let's double-check to see if this is working by setting the strength of  to 100.

Ok, refresh now! Now the Jedi Starship is stronger. Undo that last change.

So how cool is this? Not only can we have multiple objects, but they can interact with each other through methods like this one. I'll show you more of this later.