But before we go on, we need to help my editor. It's confused. Inside printShipSummary(), my editor doesn't seem to recognize the sayHello() method on Ship, it thinks it doesn't exist. But down at the bottom of the file, when I call doesGivenShipHaveMoreStrength(), it's not highlighted in yellow - that means my editor does see that this method exists. So what gives? Why doesn't it recognize the sayHello() function?
If you just look at the printShipSummary() function, all that my editor knows is that we're passing in some argument called {dollar_sign}someShip, but it doesn't know what it is. Is it a string? A boolean? A Ship object? We know that this will be a Ship object, because we're creating Ship objects below and passing those as the argument. But our editor has no idea. And for that reason, it doesn't know to look on the Ship class to see that there's a sayHello() function.