Ship class lived inside play.php, and this was just a garbage file we won't use anymore. Usually, a class will live all alone in its own file. Create a lib/ directory, and a file called Ship.php. There's no technical reason why I called the directory lib/, it just sounds nice. And the same goes for the filename - we could call it anything. But to keep my sanity, putting the Ship class inside Ship.php makes a lot more sense than putting it inside of a file called ThereIsDefinitelyNotAShipClassInHere.php. So even though nothing technical forces us to do this, put one class per file, and then go celebrate how clean things look.
Go copy the Ship class, and put it into Ship.php.

Don't put a closing PHP tag, because you don't need it. PHP will reach the end of the file, and close it automatically.

I'll head back to play.php, just like when you have functions in an external file you have to require that file to have access to it. So we'll require once __DIR__'/lib/Ship.php'.

The __DIR__ is a constant that points to this directory. So this makes sure that we're requiring exactly /lib/Ship.php relative to this file.

Get rid of requires?

If you're familiar with modern apps you'll notice that they don't have this require statement, we'll talk about that in the future. There is a way called autoloading to not even need require statements. But for now we do need it.

So now that we've moved that out let's refresh. Well look at that, it still works.