So now that we have this Ship class inside of a Ship.php file we can start using it from within our real application. From our get_ships() function, I don't want to return this array inside of an array thing anymore. I want to do awesome things like return objects.
We'll start with adding our require statement.

Next, let's transform our brave starfighter into a Ship object. We do that with  = new Ship(); and then we'll just start setting the details: name = 'Jedi Starfighter', weaponPower of 5, jediFactor of 15 and strength of 30.

The bottom array, we're not going to use that at all anymore. Instead we're going to return an array with just this one ship in it, we'll add more to our fleet later.

Remember, we're calling this back in index.php, in that file we call get_ships(); that use to return an array of ship arrays. Now it returns an array of Ship objects, of which there will only be the one starfighter. Let's var_dump this to see what it looks like.

Take off the file name, so we load index.php:

http://localhost:8000
And there it is! We have an array with one item in it which is our Ship object. Look at those sweet spaceship stats.