Treat that Object like an Object (->)
Let's take that var_dump off and see what that does to our app. When we refresh we see an exciting error that tells us we cannot use object of type Ship as array on line 68. This is an error that you might see, so let's see what's happening on line 68. This is an error that you might see, so let's see what's happening on line 68:

 107 lines  index.php
... lines 1 - 66
  <?php foreach ({dollar_sign}ships as {dollar_sign}ship): ?>
    <tr>
      <td><?php echo {dollar_sign}ship['name']; ?></td>
      <td><?php echo {dollar_sign}ship['weapon_power']; ?></td>
      <td><?php echo {dollar_sign}ship['jedi_factor']; ?></td>
      <td><?php echo {dollar_sign}ship['strength']; ?></td>
    </tr>
  <?php endforeach; ?>
... lines 75 - 107
Ok, we're using ['name']. Before when each ship was an array, that made sense, now we know when you reference an object you need to use an arrow. So if you do have an object and you try to use the square bracket syntax that is the error that you will see. Lucky you!