Back in our editor, because this is an object we can use our methods on it. In our Ship class we have this getName() method:
 50 lines  lib/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 17
    public function getName()
    {
        return {dollar_sign}this->name;
    }
... lines 22 - 48
}

Now, let's fix this little error we have in the select menu. In index.php you can see it's the same thing as before, we're using the {dollar_sign}ship like an array, so change this to use getName(); here and down there as well.

Refresh, and now things look just fine