We have this getNameAndSpecs() function, so perhaps when I'm choosing a ship I might want to see its important stats since I'm going to use it to save the day:
 50 lines  lib/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 22
    public function getNameAndSpecs({dollar_sign}useShortFormat)
    {
... lines 25 - 41
    }
... lines 43 - 48
}
So instead of getName() I'll use getNameAndSpecs(). First, I'm going to make the short format an optional argument so we don't always have to fill that in.

We see our specs format in the select menu -- cool!

And that's it, switching to an object is not that big of a deal. Next we'll talk about what these public things in Ship are doing inside of here and what else we can have.