Adding a Setter Method
This is called a visibility modifier. Once you make something private if you want someone from the outside to be able to interact with that property you'll need to add public functions to be able to do that. In this case down here, we can create what's called a setter: public function setStrength({dollar_sign}strength) it will take an argument called {dollar_sign}strength, which will be a number.

And then we'll set it on that property. You see that this does not highlight red, so a private property can still be accessed from within a class using the magic {dollar_sign}this keyword. It just can't be accessed outside of the class.

Here, instead of accessing the strength property directly, we can access the setStrength method.

When we refresh, it gets further!

Adding a Getter Method

It gets past that setter and now we're down to line 71. We're still accessing the strength property:
index.php
... lines 1 - 65
                    <?php foreach ({dollar_sign}ships as {dollar_sign}ship): ?>
... lines 67 - 70
                            <td><?php echo {dollar_sign}ship->strength; ?></td>
... line 72
                    <?php endforeach; ?>
... lines 74 - 106