Make all the Things Private!
This idea of making your properties private and then adding getters and setters is really common. Even if you don't need the control like this now you might in the future. If you're already forcing outsiders to call your setter methods and you realize later that you need to do some sort of check you have the opportunity to do that by modifying your method.

A really common thing to do is to always make your properties private. So I'll update jediFactor, weaponPower and name.

The downside of this is that we'll need a getName(), setName(), getWeaponPower(), setWeaponPower() getJediFactor(), and a setJediFactor().

That can be a lot of work and PHP doesn't give us a way to get around this, so we do need to write those. A lot of editors allow you to generate these, which is nice. In PHPStorm, go to code generate and then pick 'Getters and Setters' and select the weaponPower and jediFactor fields. Name isn't in this list because we already have a getName(). I'll go back to code generate again and pick just setter this time and it recognizes that the name doesn't have a setter.

Now that we've made everything private and we have these getters and setters, we need to use those everywhere instead of accessing the properties directly. Let's change this to setName(), this to setWeaponPower(), and setJediFactor().

Maybe this feels like extra work right now, but if we had made it private in the beginning then we wouldn't have to go back and change them. Which is what I recommend that you do.

In index.php we have the same thing, we need to call getWeaponPower() and getJediFactor().

So let's try that out and see if we missed anything. Refresh and everything looks really good and even the select menu shows up perfect. We're all set.