'Argument 1 passed to battle() must be of the type array, object given' was raised
since we're passing it a ship object. This was called a type hint and the only purpose of a type hint in PHP is to get better errors: it doesn't change the behavior. We can just take the type hint off like this and that will fix the error.

And down here, knowing that  is actually an object, instead of using the array syntax we can call the getStrength() method. Let's go ahead and dump {dollar_sign}ship1Health to make sure it's working.

This time it's printing out 60 which means it's printing out the ship's mighty strength correctly.