Type-Hinting Saves your Butt
The type hint is a useful thing, not from a functionality standpoint, but for knowing when you're doing something wrong. Let's go back to battle.php and pretend that something went wrong here by changing our object {dollar_sign}ship1 to the string foo.

When we refresh this time, we get this really weird error:

Call to a member function getStrength() on a non-object

You're going to see this error a lot and it's coming from line 76:

 129 lines  functions.php
... lines 1 - 73
function battle({dollar_sign}ship1, {dollar_sign}ship1Quantity, {dollar_sign}ship2, {dollar_sign}ship2Quantity)
{
    {dollar_sign}ship1Health = {dollar_sign}ship1->getStrength() * {dollar_sign}ship1Quantity;
    var_dump({dollar_sign}ship1Health);die;
... lines 78 - 121
}
... lines 123 - 129
It happens whenever you use the arrow syntax on something that isn't an object. It's a fatal error and PHP just dies immediately. We know because I just passed foo, that {dollar_sign}ship1 is no longer an object, it's just a string. And when we call this on it, everything dies. The issue is that from the error message, it isn't exactly clear where the mistake is. It's telling us that the problem is on line 76 in functions.php. And sure, that is where the error occurred. But the real problem is in battle.php where we are passing in a bad value to the battle() function.