Type-Hinting with a Class
So in addition to type-hinting with the array, when we use objects we can type hint with the class name. This means we can actually type Ship.

That is the exact same thing. It says, 'Hey, PHP, if something is passed to this argument that's not a Ship objec, I want you to throw a very clear error.' So let's go see this new error! Refresh and there it is:

Argument 1 passed to battle() must be an instance of Ship, string givenon line 29

This time it's very clear: it says it should have been a Ship object, but you're passing a string and it points us to the exact right spot. So type-hinting is optional, but it's a really good idea because it will make your code easier to debug later. It also has a second benefit: as soon as I type hinted this {dollar sign}ship1 variable here, all of a sudden my editor knew what type of object  was and offered me autocomplete. So it knows about getStrength() and all the other methods on that object.