Here's the next challenge, sometimes I want my ships to be broken. So when a ship comes in it might be able to fight or it might be under repair. This is a property on a Ship, a Ship could be under repair or not. So I'll add this as a new private property that has a boolean which can be true or false.
The challenge is that whenever we create our objects inside of functions.php, I want to randomly set a Ship to be under repair or not under repair, and I want it to happen automatically. So just by creating a Ship, I want it to internally figure out if it is under repair or not.