public function __construct
This is where the idea of a constructor comes in. Whenever you create an object, you can actually hook into that process and say: 'Hey whenever a Ship is created, call this function so I can set some stuff up.' The way you do this is by creating a very special public function inside of your class called __construct().

The magic here is that name: it must be __construct. And just by having this it should be called everytime we say new Ship().

Let's try it, refresh! And that's it: it's called four times, once for each of our ships. And it's called right when you say new Ship(), so if I throw in a die() statement right after creating a Ship, we're still going to see one of those called.