Setting up Data in __construct
Now we have a really powerful way to set up our data. Internally we can determine whether or not the Ship is under repair.

This gives each ship a 30% chance of being broken...maybe a wing fell off!

To see this in action, let's cheat real quick and var_dump the {dollar_sign}ships array. When we refresh we can see the first two ships are ready for action but the third one isn't. Refresh again and they're all ready to fly. And a third time shows that the first two are busted and the other two are battle worthy. So that's working already.