Don't Create a Getter Function
Next, let's go into index.php and up top we have our table information. Let's include status which will tell us if our ship is under repair or not.

Now so far, we don't have a way to access that new property. It's private and we don't have a getter or a setter and you don't necessarily need to create these. In fact, we don't want a setter: it's being set automatically inside of the class itself. But I do want to figure out if this Ship is functional or not. So what I'll do is create a new public function and I'll call it isFunctional().

This will be the opposite of the underRepair value. If it is underRepair, then it is not functional and if it is functional then it is not underRepair. For the outsider whose going to be calling this function, they don't care what we're doing internally to figure that out.

A sunshine for functional and a sad cloud for not functional.

Refresh and try it out, four sunshines and one cloud.