Adding Arguments to __construct
The __construct() function is something you are going to see a lot but it's a really easy idea. It just says if you have a function called __construct(), then it's automatically going to be called when you instantiate your object.

There is one other thing you can do: like most functions, it can have an argument. Let's put a name argument.

I'm not going to use it yet because I'm going to show you what happens when we do that.

Go back to functions.php. You can see that my editor is angry because it says required parameter  missing.

So, you notice whenever we create a new Ship object, it's always Ship(), but you never pass anything in there. When you create an object, the stuff that goes in between the parenthesis are arguments that are passed to your __construct() function, if you have one. Because we have a  argument here, now we need to pass a name there.

Now you can see that it is happy.

And what we can do inside of Ship.php is say, ok whatever {dollar_sign}name they pass in, let's just set that to the name property.

When to Pass Values to __construct

So, why would you do this? Why would you add a  argument to the Ship's constructor and force it to be passed in versus the setter? It's really up to you. In our case, it doesn't make sense to have a Ship without a name. And before, that would have been possible had we just instantiated a new Ship and forgotten to call setName(). Then we would have been running around with a Ship object that had absolutely no name. How embarrasing.

Sometimes, when you have required information, you might choose to set them up as arguments to your constructor. It says 'Hey, when you create a Ship, you have to pass me a name.' We're not forcing the user to pass a weaponPower, jediFactor or strength, because those all have a nice default value of 0. So it makes sense not to force those, but we do force the name.

When you back up I just want you to realize that the __construct function is just like any other function. But if you give it that special name, it is automatically called and the arguments are passed to it.

And guess what! You just learned the fundamentals of object-oriented programming. Classes, check! Objects, super check! Methods, privacy, type-hinting, constructor and other stuff - all old news now. And there's even more great stuff to learn, like service classes, dependency injection, inheritance and interfaces. These will make you even more dangerous, and will also help you understand the outside libraries you use everyday.